third_party/blink/web_tests/fast/table/caption-in-inline-expected.txt

Test that a table-caption with a table-cell sibling is wrapped inside an inline-table when the parent is inline.

 
 There should be a blue square to the left of this line.
PASS firstRects.length is 1
PASS secondRects.length is 1
PASS lastRects.length is 1
PASS secondRects[0].top is firstRects[0].top + 40
PASS secondRects[0].left is firstRects[0].left
PASS secondRects[0].bottom is secondRects[0].top + 40
PASS lastRects[0].right is >= secondRects[0].left
PASS lastRects[0].top is >= firstRects[0].top
PASS secondRects[0].bottom is >= lastRects[0].bottom
PASS successfullyParsed is true

TEST COMPLETE
